ITSM/requirements.txt
# pip install mysql-connector-python tabulate

mysql-connector-python
tabulate